requirements.txt
# ollama
streamlit
langchain_core
langchain_community
langchain_ollama

# vector
faiss-cpu
sentence-transformers
numpy


matplotlib
scikit-learn

# tts
coqui-tts